feat: calculation process rework
Designed and implemented an algorithm of filtering strings,
used it to finish the calculation pipeline.

Now every hamiltonian group is further divided into colors, and
every color represents a set of pairwise commutative terms in the
hamiltonian. We can determine which of the terms belonging to a certain color
act on the specific site, that helps us to filter out unnecessary
multipliers during the calculation.